requirements.txt
asgiref==3.8.1
astroid==3.3.8
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.3.9
Django==5.1.5
django-crispy-forms==2.3
django-import-export==4.3.4
django-stubs==5.1.2
django-stubs-ext==5.1.2
et_xmlfile==2.0.0
isort==5.13.2
joblib==1.4.2
MarkupPy==1.14
mccabe==0.7.0
numpy==2.2.2
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pylint==3.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sklearn==0.0
sqlparse==0.5.3
tablib==3.8.0
threadpoolctl==3.5.0
tomlkit==0.13.2
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
xlrd==2.0.1
xlwt==1.3.0
